collectd: Add an API to scollectd data
The scollectd is an infrastructure that allows different part of the
code to register internal counters and the infrastructure would send it
periodically to an external server.

This patch adds and API to the scollectd that allows to inquire a
register value and the names of the registered values.

The collectd_value structure is used to return a single value that can
be of type: double or signed and unsigned 64 bit long.

The definition of the API are found in scollectd_api.hh

The inquiries are for the local cpu, it is up to the caller to call a
relevent cpu.

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>