gossip: Fix sending ECHO message
The msg parameter is missing.

Fix a bug where Node B does not not recognize Node A.

Node A
$ ./gossip --seed 127.0.0.1  --listen-address 127.0.0.1

Node B
$ ./gossip --seed 127.0.0.1  --listen-address 127.0.0.2

The issue is that in gossiper::mark_alive(), the parameter for ECHO
message is wrong and after commit 1a8c4b75f5 (message: do not erase client's
rpc call type), we deduce the rpc handler using the parameters supplied
to messaging_service::send_message(), so we will use a wrong handler for
the ECHO message and the message will never reply thus we never mark the
peer node alive.

empty_msg was used as a placeholder when messaging_service does not handle void
return type correctly. Since we support it now, drop it.