sstables: change read_simple and write_simple to be able to handle arbitrary parameters
Passing all arguments as template parameters was a nice trick, made possible by
the fact that all fields we were dealing with were part of the sstable
structure.

However, as we progress, it would make sense for some fields to be short-lived.
In that case, the that convenience turns into an inconvenience.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>