rpc: add counters to the rpc client
This adds a counters to net protocol client.
The client would holds counter for the successfully completed, the
exception recieved and as the sent queue is based on the future queue
a counter for pending to sent messages.

The number of messages that waiting for reply is retrieved from the size
of the outstanding table.

The counters are updated by the helper send function as part of the
messages sending flow.

The counters are part of the stat structure and the client has an
internal getter to be able to change their values and a getter that
returns a copy of the counters.