bloom filter: fix incorrect size
There is a tricky bug in our current filter implementation: is_present will
return a different value depending on the order keys are inserted.

The problem here, is that _bitmap.size() will return the maximum *currently*
used bit in the set. Therefore, when we hash a given key, the maximum bit it
sets in the bitmap is used as "max" in the expression

    results[i] = (base % max);

If the next keys do not set any bit higher than this one, everything works as
expected, because the keys will always hash the same way.

However, if one of the following keys happens to set a bit higher than the
highest bit which was set at the time a certain key was set, it will hash using
two different values of "max" in the aforementioned expression; one at insertion,
and another one at the test.

We should be using max_size() to be sure that we will always have the same
hash results.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>