sstables: get rid of write_static_column_name
After the last round of cleanups, this function turned out to be exactly the
same as write_column_name, except that the composite differs. Because the
composite is passed as a parameter, we can just use the same function for all
and pre-create the composite.

This will make the implementation of collections a lot easier, since for
collections we will prepend each element with the column name.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>