sstables: convert EstimatedHistogram to C++ and start using it
In addition, this patch also fixes serialization and deserialization of
estimated histogram. Problem was found by reading the respective methods
in origin implementation.

The first element of the array offset is used for both the first and
second element of the array bucket. So given an array bucket of size N,
array offset will be of size N - 1. Our code wasn't handling this.

The new representation of estimated histogram provides us with methods
needed for writing the component Statistics.

Signed-off-by: Raphael S. Carvalho <raphaelsc@cloudius-systems.com>