convert sstable write path to seastar threads
This is my attempt to convert the sstable code to seastar threads, as
we have been extensively discussing. I haven't yet measured how much do we
gain by this, but the final code looks *so* much better and less complicated,
that this alone should be enough reasoning.

Here's how I've done it, so you can easily follow:

Every function that we use and returns a future, is copied to another function
with the same name but ending in _t. This is better than copying the whole thing,
because it can be done in logical pieces that are easier to follow. This is also
easier to verify.

function_t() will do the same as function(), but will return void.

I am not changing more than I need to, so in the final code, without all the
do_withs and other stuff, there are some parts that start to cry for a cleanup.
They are left as is for now, and I will return to them later once the patch is
merged.

In this initial patch, you get the main write_components converted, and this
nice explanation message.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>