write all the components
checksum and crc are written inside the main function so we don't need
to export the file stream. But since the functions are actually trivial
we can just .get() the whole thing instead of changing them.

The others are still kept as futures and called after async::thread completes,
for maximum parallelism.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>