correctly position end of row
In a testament to how confusing our old code was, while collapsing the futures
I ended up getting the end_of_row element inside the loop for clustered keys.
The end of row, as the name implies, should only be written at the end of the
(thrift) row (= CQL partition).

Move it to the right place.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>