build: strip all tests
[I'm sending this patch to Urchin, but it will also apply (with different
description, to Seastar]

This patch saves about 12 GB of disk space in Urchin's build directory.

Our debugging information is absurdly huge, and added to every single
executable we generate. This is most noticable in tests - we currently
have about 60 test executables, and each one contains a copy of all
this debug information.

This patch creates all test executables - whether debug or release mode -
stripped. When a user encounters a failing test he wants debug information
for (for gdb or the sanitizer), he can trivially relink it unstripped,
with a command like:

	ninja build/debug/tests/urchin/sstable_test_g

(note the added "_g"). This links the already existing object files
(which still have their debug information), which takes a few seconds.

On my machine, this patch reduces the Urchin build directory from about
22 GB to less than 10 GB. The build/release/tests directory drops from
6.8 GB to just 0.6 GB! The build/debug/tests directory is still huge (6 GB)
and could be improved further, but is also less than half of what it used
to be.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>