sstables: separate summary creation in two parts
We do a lot more work than we should in prepare_summary. In particular,
in cases where we will create sstables from compression, some of that information
will not be known until later.

This code is also fragile, as exposed by a recent bug, in which we miscalculated
s.header.size, ending up with a completely unusable file.

The scheme proposed in this patch makes sure things like that does not happen,
because we will be generating the header based on what we actually write, and
not based on what we estimate that we will write.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>