sstables: correctly compute number of entries in summary
We are always adding one to the number of entries, but that is incorrect.  If
we have an exact multiple of the sampling rate, then we shouldn't.

This is no longer affecting us as a bug, because we now compute the size after
we seal the file. However, we should still do it right, because we will need
the correct value at some point, if we support changes to the sampling level.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>