database: be more forgiving with sstables parsing
Currently, Origin generates sstables in the form CF-UUID, where UUID
is a string of numbers.

We also do CF-UUID, but for us, UUID has dashes separating the UUID components.

Due to the current test, we fails to load our current sstables. That test
really isn't that important, since we are currently not doing anything with the
UUID. And if we were, we should be able to accept both formats anyway.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>