sstables: make sure sstable directory exists
In theory, when we create a new column family, we should also make sure
that the underlying directory exist. However, this would be quite challenging:
there are a lot of entry points for, add_column_family, none of them are futurized,
and futurizing them could prove challenging up the call chain.

Because we can guarantee that the keyspace directory will exist - now that we
have unified that, it is actually a lot simpler to just make sure that the
directory exist when writing the sstable.

If the keyspace directory wouldn't exist we would have to recurse through the
path.  As previously said, this patch will assume this away.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>